+++ shop_functions.inc.php
@@ -173,6 +173,7 @@ function ProductList($cat_id = 0)
         !empty($_REQUEST['query']) &&
         !isset($_REQUEST['clearsearch'])
     ) {
+        $query_str = urlencode($_REQUEST['query']);
         $search = DB_escapeString($_REQUEST['query']);
         $fields = array('p.name', 'c.cat_name', 'p.short_description', 'p.description',
                 'p.keywords');
@@ -246,12 +247,14 @@ function ProductList($cat_id = 0)
         'pi_url'        => SHOP_URL,
         //'user_id'       => $_USER['uid'],
         'currency'      => $_SHOP_CONF['currency'],
-        'breadcrumbs'   => $cat_id > 0 ? Category::Breadcrumbs($cat_id) : '',
+        'breadcrumbs'   => $cat_id > 0 ? $Cat->Breadcrumbs() : '',
         'search_text'   => $search,
         'tpl_ver'       => $_SHOP_CONF['list_tpl_ver'],
         'sortby_options' => $sortby_options,
         'sortby'        => $sortby,
         'table_columns' => $_SHOP_CONF['catalog_columns'],
+        'cat_id'        => $cat_id == 0 ? '' : $cat_id,
+        'query'         => $query_str,
     ) );
 
     if (!empty($cat_name)) {